Moving java2sdg exceptions in a separate error folder
