+++ seep-java2sdg/src/main/java/uk/ac/imperial/lsds/java2sdg/errors/LimitationException.java
@@ -1,4 +1,4 @@
-package uk.ac.imperial.lsds.java2sdg;
+package uk.ac.imperial.lsds.java2sdg.errors;
 
 import uk.ac.imperial.lsds.seep.errors.SeepException;
